fix(Route.h): Revert some inadvertent Route.h changes
The inclusion of several methods in the new Route.h was inadvertent. IsAtEnd, NextStop, Get* are NOT necessary. It is part of the solution version, but does NOT need to be used by you in your implementation.  Similarly the two added data members destination_stop_index_ and destination_stop_ were used in the solution and do not need to be in your solution.  They are now commented out. Use them if you wish. Ignore them if you like. Sorry for any confusion.