Update page length spec
Change from prior text (in an attempt to clarify) made the requirement 3-4 for the entire mainpage. The 3-4 page estimate should be applied to the User guide. Given an introduction, instructions on download and configuration, instructions on running both version (including probably some images), and discussion of the config files (and how to make more), 3-4 pages should be reasonable.